Update Pca95x4 project
